+++ x/domain/types/types.go
@@ -20,14 +20,21 @@ type Domain struct {
 	Admin sdk.AccAddress `json:"admin"`
 	// ValidUntil is a unix timestamp defines the time when the domain will become invalid
 	ValidUntil int64 `json:"valid_until"`
-	// HasSuperuser checks if the domain is owned by a super user or not
-	HasSuperuser bool `json:"has_super_user"`
+	// Type defines the type of the domain
+	Type DomainType `json:"type"`
 	// AccountRenew defines the duration of each created or renewed account
 	// under the domain
 	AccountRenew time.Duration `json:"account_renew"`
 	// Broker TODO needs comment
 	Broker sdk.AccAddress `json:"broker"`
 }
+
+type DomainType string
+
+const (
+	OpenDomain  DomainType = "open"
+	CloseDomain            = "close"
+)
 
 // Index implements Indexer and packs the
 // domain into an index key using its name